Modulize STT code
